test diversa costruzione della parte MVC di offerta

+++ src/main/java/com/corsojava/springcrud/repository/OffertaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.corsojava.springcrud.model.Offerta;
 
-public interface OffertaRepository extends JpaRepository<Offerta, Long> {
+public interface OffertaRepository extends JpaRepository<Offerta, Integer> {
 
 }
